modules/location_consumer/requirements.txt
aniso8601==9.0.1
attrs==21.4.0
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
GeoAlchemy2==0.12.1
greenlet==1.1.2
grpcio==1.46.3
grpcio-tools==1.46.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.6.0
kafka-python==2.0.2
MarkupSafe==2.1.1
packaging==21.3
protobuf==3.20.1
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.1
Shapely==1.8.2
six==1.16.0
SQLAlchemy==1.4.37
typing_extensions==4.2.0
Werkzeug==2.1.2
zipp==3.8.0
